index: track cache_offset in map
It can change if the message gets archived underneath us.  The alternative
is to read that system-flag from the disk, but then you get into the
horrible land of write races, and partial reads, which locking protects
us from.  Better to just spend the extra 4 bytes

(which rounds up to 8 - but we can get them ALL back if we're willing
to play bitmap games with the empty space in system_flags)