archive partition support
This involves massive changes to caching, using mappedfile.  It
involves new config options.  There's a bugzilla bug, which should
be mentioned in here.

For FastMail, we intend to hand the new messages out on SSD, and then
shunt all the old ones off onto slow storage.  It cuts the fsync count
to the slow disks by a LOT.